manager_ai/requirements.txt
requests>=2.28.0
requests-oauthlib>=1.3.0
python-dotenv>=1.0.0
rich>=13.0.0
flask>=3.0.0
